requirements.txt
PySimpleGUI
PyCryptodome
pillow
